requirements.txt
python-telegram-bot==20.7
groq==0.4.2
nest-asyncio==1.5.8
Pillow==10.1.0
ollama==0.1.7
python-dotenv==1.0.0 